requirements.txt
colorama==0.4.4
python-nmap==0.7.1
requests==2.26.0
